# openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerIn": {
        "properties": {
          "chosen_index": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Index"
          },
          "essay_text": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Essay Text"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          }
        },
        "required": [
          "question_id"
        ],
        "title": "AnswerIn",
        "type": "object"
      },
      "AnswerSavedOut": {
        "properties": {
          "answered_count": {
            "title": "Answered Count",
            "type": "integer"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "question_id",
          "answered_count",
          "total_count"
        ],
        "title": "AnswerSavedOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfflineQuestionOut": {
        "properties": {
          "answer_hmac": {
            "title": "Answer Hmac",
            "type": "string"
          },
          "difficulty": {
            "title": "Difficulty",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          },
          "qtype": {
            "title": "Qtype",
            "type": "string"
          },
          "source_excerpt": {
            "title": "Source Excerpt",
            "type": "string"
          },
          "stem": {
            "title": "Stem",
            "type": "string"
          }
        },
        "required": [
          "id",
          "qtype",
          "stem",
          "options",
          "difficulty",
          "source_excerpt",
          "answer_hmac"
        ],
        "title": "OfflineQuestionOut",
        "type": "object"
      },
...
      "QuizOut": {
        "properties": {
          "answered_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Answered Ids",
            "type": "array"
          },
          "base_reward_seconds": {
            "title": "Base Reward Seconds",
            "type": "integer"
          },
          "level_factor": {
            "title": "Level Factor",
            "type": "number"
          },
          "material_id": {
            "format": "uuid",
            "title": "Material Id",
            "type": "string"
          },
          "max_reward_seconds": {
            "title": "Max Reward Seconds",
            "type": "integer"
          },
          "novelty_factor": {
            "title": "Novelty Factor",
            "type": "number"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionPublic"
            },
            "title": "Questions",
            "type": "array"
          },
          "ready_count": {
            "title": "Ready Count",
            "type": "integer"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "material_id",
          "title",
          "base_reward_seconds",
          "level_factor",
          "novelty_factor",
          "max_reward_seconds",
          "ready_count",
          "total_count",
          "status",
          "questions"
        ],
        "title": "QuizOut",
        "type": "object"
      },
      "QuizStartIn": {
        "properties": {
          "material_id": {
            "format": "uuid",
            "title": "Material Id",
            "type": "string"
          }
        },
        "required": [
          "material_id"
        ],
        "title": "QuizStartIn",
        "type": "object"
      },
      "ReceiptOut": {
        "properties": {
          "balance_seconds": {
            "title": "Balance Seconds",
            "type": "integer"
          },
          "base_reward_seconds": {
            "title": "Base Reward Seconds",
            "type": "integer"
          },
          "correct_count": {
            "title": "Correct Count",
            "type": "integer"
          },
          "credited_seconds": {
            "title": "Credited Seconds",
            "type": "integer"
          },
          "gross_seconds": {
            "title": "Gross Seconds",
            "type": "number"
          },
          "level_factor": {
            "title": "Level Factor",
            "type": "number"
          },
          "level_note": {
            "title": "Level Note",
            "type": "string"
          },
          "new_badges": {
            "items": {
              "type": "string"
            },
            "title": "New Badges",
            "type": "array"
          },
          "novelty_factor": {
            "title": "Novelty Factor",
            "type": "number"
          },
          "novelty_note": {
            "title": "Novelty Note",
            "type": "string"
          },
          "question_count": {
            "title": "Question Count",
            "type": "integer"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ReceiptRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "streak_current": {
            "title": "Streak Current",
            "type": "integer"
          },
          "subtotal_seconds": {
            "title": "Subtotal Seconds",
            "type": "number"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "session_id",
          "title",
          "base_reward_seconds",
          "rows",
          "subtotal_seconds",
          "level_factor",
          "level_note",
          "novelty_factor",
          "novelty_note",
          "gross_seconds",
          "credited_seconds",
          "balance_seconds",
          "correct_count",
          "question_count",
          "streak_current",
          "new_badges"
        ],
        "title": "ReceiptOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Kartu waktu bermain yang dicetak dari materi belajar",
    "title": "BrainXP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/quizzes/{session_id}": {
      "get": {
        "operationId": "get_quiz_quizzes__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quiz",
        "tags": [
          "quiz"
        ]
      }
    },
    "/quizzes/{session_id}/answers": {
      "post": {
        "operationId": "answer_quizzes__session_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerSavedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer",
        "tags": [
          "quiz"
        ]
      }
    },
    "/quizzes/{session_id}/submit": {
      "post": {
        "operationId": "submit_quizzes__session_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit",
        "tags": [
          "quiz"
        ]
      }
    },
...
    "/subjects/{subject_id}/offline-bank": {
      "get": {
        "operationId": "offline_bank_subjects__subject_id__offline_bank_get",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OfflineQuestionOut"
                  },
                  "title": "Response Offline Bank Subjects  Subject Id  Offline Bank Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Offline Bank",
        "tags": [
          "quiz"
        ]
      }
    },
...
    "/subjects/{subject_id}/quizzes": {
      "post": {
        "operationId": "start_quiz_subjects__subject_id__quizzes_post",
        "parameters": [
          {
            "in": "path",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Quiz",
        "tags": [
          "quiz"
        ]
      }
    },
...
  }
}